server/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
protobuf==4.25.2
requests==2.31.0
urllib3==2.1.0
